feat: adiciona menu interativo de console e corrige alguns erros

+++ src/main/java/exceptions/InsuficientBalanceException.java
@@ -1,7 +1,7 @@
 package exceptions;
 
 public class InsuficientBalanceException extends Exception {
-    public InsuficientBalanceException(int value) {
+    public InsuficientBalanceException(double value) {
         super("Saldo insuficiente. Saldo disponível: R$: " + value + ".");
     }
 }
